+++ src/index.d.ts
@@ -158,9 +158,17 @@ interface NumberAdjuster extends AdjusterBase<number>
 	maxValue(value: number, adjust?: boolean): this
 
 	/**
+	 * conversion
+	 * @param converter conversion function
+	 * @returns chainable instance
+	 */
+	convert(converter: (value: number, fail: () => never) => number): this
+
+	/**
 	 * mapping
 	 * @param mapper mapping function
 	 * @returns chainable instance
+	 * @deprecated use convert()
 	 */
 	map(mapper: (value: number, fail: () => never) => number): this
 }
@@ -230,9 +238,17 @@ interface StringAdjuster extends AdjusterBase<string>
 	pattern(pattern: RegExp): this
 
 	/**
+	 * conversion
+	 * @param converter conversion function
+	 * @returns chainable instance
+	 */
+	convert(converter: (value: string, fail: () => never) => string): this
+
+	/**
 	 * mapping
 	 * @param mapper mapping function
 	 * @returns chainable instance
+	 * @deprecated use convert()
 	 */
 	map(mapper: (value: string, fail: () => never) => string): this
 }
@@ -302,9 +318,17 @@ interface NumericStringAdjuster extends AdjusterBase<string>
 	checksum(algorithm: string): this
 
 	/**
+	 * conversion
+	 * @param converter conversion function
+	 * @returns chainable instance
+	 */
+	convert(converter: (value: string, fail: () => never) => string): this
+
+	/**
 	 * mapping
 	 * @param mapper mapping function
 	 * @returns chainable instance
+	 * @deprecated use convert()
 	 */
 	map(mapper: (value: string, fail: () => never) => string): this
 }
@@ -451,6 +475,8 @@ type ConstantsCause = {
 	NULL: string,
 	EMPTY: string,
 	ONLY: string,
+	MAP: string, // deprecated; use CONVERT
+	CONVERT: string,
 
 	MIN_VALUE: string,
 	MAX_VALUE: string,